[refactor] Restructure pipeline for performance and maintainability
- Split monolithic 400+ line pipeline into modular ETL structure
- Replace pandas with polars for 4+ minute processing optimization
- Separate data extraction/transformation from SQL-based analytical modeling
- Create materialization layer with SQL templates and DuckDB execution
- Move from print() to proper logging with YAML configuration
- Organize data into structured marts/models/raw/processed directories
- Update CI workflows and app paths for new architecture
