--- .github/ISSUE_TEMPLATE/config.yml
@@ -1,4 +1,0 @@
-contact_links:
-    - name: 🤔 Questions and Suggestions
-      url: mailto:gaelop101@protonmail.com
-      about: Please file questions and suggestions here.
