--- .github/workflows/model.yml
@@ -1,45 +1,0 @@
-name: Model Data
-
-on: 
-  workflow_dispatch:
-  workflow_run:
-    workflows: ["Import Data"]
-    types: 
-      - completed
-
-jobs:  
-  modeling:
-    runs-on: ubuntu-latest
-    if: ${{ github.event.workflow_run.conclusion == 'success' }}
-    steps:
-
-      - name: checkout repo content
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
-
-      - name: setup python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.10'
-
-      - name: install python packages
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-      
-      - name: execute py script
-        run: python modeling.py
-
-      - name: commit files
-        run: |
-          git config --local user.email "action@github.com"
-          git config --local user.name "GitHub Action"
-          git add -A
-          git diff-index --quiet HEAD || (git commit -a -m "Update models" --allow-empty)
-
-      - name: push changes
-        uses: ad-m/github-push-action@v0.6.0
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          branch: master
