+++ .github/workflows/update.yml
@@ -1,0 +1,110 @@
+name: Update Data
+
+on: 
+  workflow_dispatch:
+  schedule:
+    - cron: '0 2 1 */1 *'
+
+jobs:
+  download-json:
+    runs-on: ubuntu-latest
+    steps:
+      - name: checkout repo content
+        uses: actions/checkout@v4.1.7
+
+      - name: pull changes from remote
+        run: git pull origin
+
+      - name: check repository size
+        run: du -h -d 0
+      
+      - name: download json file
+        run: |
+          wget --secure-protocol=TLSv1_2 -c --retry-connrefused --waitretry=1 --read-timeout=20 --timeout=15 -t 0 --continue -O "data/raw/weapons-wanted.json" "https://data.gov.ua/dataset/d0af9ba0-08b3-4bca-8508-02cffeaae8fd/resource/1fcab772-0b3c-4938-8f72-e60db343cbe5/download/weaponswanted.json"
+
+      - name: upload json for pipeline job
+        uses: actions/upload-artifact@v4
+        with:
+          name: weapons-wanted
+          path: data/raw/weapons-wanted.json
+          retention-days: 1
+      
+
+  run-etl:
+    runs-on: ubuntu-latest
+    needs: download-json
+    steps:
+      steps:
+      - name: checkout repo content
+        uses: actions/checkout@v4.1.7
+
+      - name: pull changes from remote
+        run: git pull origin
+
+      - name: setup python
+        uses: actions/setup-python@v5.1.0
+        with:
+          python-version: '3.13'
+
+      - name: install polars library
+        run: |
+          python -m pip install --upgrade pip
+          pip install polars
+
+      - name: download json artifact for pipeline job
+        uses: actions/download-artifact@v4
+        with:
+          name: weapons-wanted
+          path: data/raw
+
+      - name: run orchestrator script
+        run: pipeline/etl.py
+
+      - name: check repository size
+        run: du -h -d 0
+        
+      - name: commit files
+        run: |
+          git config --local user.email "action@github.com"
+          git config --local user.name "GitHub Action"
+          git add -A
+          git diff-index --quiet HEAD || (git commit -a -m "[update] Update main dataset" --allow-empty)
+          git push
+        env:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+
+
+  run-modeling:
+    runs-on: ubuntu-latest
+    needs: run-etl
+    if: success()
+
+    steps:
+      - name: checkout repo content
+        uses: actions/checkout@v4.1.7
+
+      - name: pull changes from remote
+        run: git pull origin
+
+      - name: setup python
+        uses: actions/setup-python@v5.1.0
+        with:
+          python-version: '3.13'
+
+      - name: install duckdb library
+        run: |
+          python -m pip install --upgrade pip
+          pip install duckdb
+      
+      - name: execute python script
+        run: python pipeline/etl/materialize.py
+
+      - name: commit and push changes
+        run: |
+          git config --local user.email "action@github.com"
+          git config --local user.name "GitHub Action"
+          git add -A
+          git diff-index --quiet HEAD || (git commit -m "[update] Update models" --allow-empty)
+          git push
+        env:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
