+++ data/models/date-report-total.sql
@@ -1,0 +1,11 @@
+SELECT
+    date,
+    report,
+    COUNT(report) AS total
+FROM 
+    read_parquet('{processed_path}')
+GROUP BY
+    1,2
+ORDER BY
+    date ASC,
+    report ASC;
