+++ mypy.ini
@@ -1,0 +1,15 @@
+[mypy]
+python_version = 3.10.13
+strict = True
+warn_unused_ignores = True
+no_implicit_optional = True
+disable_error_code = no-any-return
+
+[mypy-polars.*]
+ignore_missing_imports = True
+
+[mypy-config.*]
+ignore_missing_imports = True
+
+[mypy-etl.*]
+ignore_missing_imports = True
