+++ pipeline/__init__.py
@@ -1,0 +1,1 @@
+# Marks pipeline/ as a package.
